Added apk with fix change.
The only other change was that the actionbarsherlock library was removed due to incompatibilities, an alternative can be considered if necessary (the only change is that now there is no action bar present, but wasn't used anyway).

If you want to use it remember to uninstall first the original one. Keep in mind that future updates of the original app won't be detected by play store, a manual update will be required (if possible I'll notify it from this github).